Rename base package to org.kiwiproject
* Rename base package from com.smoketurner to org.kiwiproject
* Update the POMs (except for now keeping parent POM of root POM)
* Update README dependency examples, and use placeholder version

Closes #12
