--- consul-core/src/main/java/com/smoketurner/dropwizard/consul/ConsulConfiguration.java
@@ -1,8 +1,0 @@
-package com.smoketurner.dropwizard.consul;
-
-import io.dropwizard.Configuration;
-
-@FunctionalInterface
-public interface ConsulConfiguration<C extends Configuration> {
-  ConsulFactory getConsulFactory(C configuration);
-}
