+++ consul-core/src/main/java/org/kiwiproject/dropwizard/consul/ConsulConfiguration.java
@@ -1,0 +1,8 @@
+package org.kiwiproject.dropwizard.consul;
+
+import io.dropwizard.Configuration;
+
+@FunctionalInterface
+public interface ConsulConfiguration<C extends Configuration> {
+  ConsulFactory getConsulFactory(C configuration);
+}
