+++ consul-example/pom.xml
@@ -4,9 +4,9 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.smoketurner.dropwizard</groupId>
+        <groupId>org.kiwiproject</groupId>
         <artifactId>dropwizard-consul</artifactId>
-        <version>2.0.12-2-SNAPSHOT</version>
+        <version>0.5.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>consul-example</artifactId>
