+++ consul-example/src/main/java/com/example/helloworld/HelloWorldConfiguration.java
@@ -1,8 +1,10 @@
 package com.example.helloworld;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.smoketurner.dropwizard.consul.ConsulFactory;
-import com.smoketurner.dropwizard.consul.ribbon.RibbonJerseyClientConfiguration;
+
+import org.kiwiproject.dropwizard.consul.ConsulFactory;
+import org.smoketurner.dropwizard.consul.ribbon.RibbonJerseyClientConfiguration;
+
 import io.dropwizard.Configuration;
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
