+++ consul-ribbon/src/main/java/org/smoketurner/dropwizard/consul/ribbon/ConsulServiceDiscoverer.java
@@ -1,0 +1,10 @@
+package org.smoketurner.dropwizard.consul.ribbon;
+
+import com.orbitz.consul.Consul;
+import com.orbitz.consul.model.health.ServiceHealth;
+import java.util.Collection;
+
+@FunctionalInterface
+public interface ConsulServiceDiscoverer {
+  Collection<ServiceHealth> discover(Consul consul);
+}
