+++ pom.xml
@@ -8,8 +8,9 @@
         <version>2.1.4-3</version>
     </parent>
 
+    <groupId>org.kiwiproject</groupId>
     <artifactId>dropwizard-consul</artifactId>
-    <version>2.0.12-2-SNAPSHOT</version>
+    <version>0.5.0-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <name>dropwizard-consul</name>
